requirements.txt
asgiref==3.11.1
colorama==0.4.6
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
pytest-html==4.2.0
pytest-metadata==3.1.1
sqlparse==0.5.5
tzdata==2025.3
